Handle some benchmark conversion warns

+++ ref_app/src/app/benchmark/app_benchmark_pi_spigot.cpp
@@ -64,8 +64,8 @@ auto app::benchmark::run_pi_spigot() -> bool
         app_benchmark_pi_spigot_digits10 < app_benchmark_pi_spigot_object.get_output_digit_count();
       ++app_benchmark_pi_spigot_digits10)
   {
-    const std::uint8_t by_calculated = app_benchmark_pi_spigot_out_.at(app_benchmark_pi_spigot_digits10);
-    const std::uint8_t by_control    = app_benchmark_pi_spigot_ctrl.at(app_benchmark_pi_spigot_digits10);
+    const std::uint8_t by_calculated = app_benchmark_pi_spigot_out_.at(static_cast<std::size_t>(app_benchmark_pi_spigot_digits10));
+    const std::uint8_t by_control    = app_benchmark_pi_spigot_ctrl.at(static_cast<std::size_t>(app_benchmark_pi_spigot_digits10));
 
     const bool next_digit_is_ok = (by_calculated == by_control);
